refactor(structure): move Lotto class to models directory
